Extract ContentShell component
